requirements.txt

flask>=2.0.0
scikit-learn>=1.0.0
numpy>=1.20.0
joblib>=1.0.0
google-cloud-aiplatform>=1.0.0